ui: redesign permission prompt as composer-slot card
Replaces the amber toast docked above the timeline with a minimal card
that takes over the composer slot while a permission request is pending,
mirroring the AskUserQuestion pattern. Drops the per-action outline
buttons and hardcoded amber/emerald/violet hues in favor of brand
semantic tokens (card / muted / accent / foreground).
